multimedia: add toolbar
